feat: Add file information to the onDocumentLoad event
