+++ packages/core/src/types/ZoomEvent.ts
@@ -1,0 +1,14 @@
+/**
+ * A React component to view a PDF document
+ *
+ * @see https://react-pdf-viewer.dev
+ * @license https://react-pdf-viewer.dev/license
+ * @copyright 2019-2021 Nguyen Huu Phuoc <me@phuoc.ng>
+ */
+
+import type { PdfJs } from './PdfJs';
+
+export interface ZoomEvent {
+    doc: PdfJs.PdfDocument;
+    scale: number;
+}
